+++ src/app/Components/ComponentsSamples/Counters/CountersSamplesViewModel.cs
@@ -1,0 +1,50 @@
+using DIPS.Mobile.UI.Components.Counters;
+using DIPS.Mobile.UI.MVVM;
+
+namespace Components.ComponentsSamples.Counters;
+
+public class CountersSamplesViewModel : ViewModel
+{
+    private int m_value = 5;
+    private int m_secondaryValue;
+    private CounterDisplayMode m_mode = CounterDisplayMode.Single;
+    private bool m_isUrgent;
+    private bool m_isSecondaryUrgent;
+
+    public List<CounterDisplayMode> Modes { get; } =
+    [
+        CounterDisplayMode.Single,
+        CounterDisplayMode.Double,
+        CounterDisplayMode.Auto
+    ];
+    
+    public int Value
+    {
+        get => m_value;
+        set => RaiseWhenSet(ref m_value, value);
+    }
+
+    public int SecondaryValue
+    {
+        get => m_secondaryValue;
+        set => RaiseWhenSet(ref m_secondaryValue, value);
+    }
+
+    public bool IsUrgent
+    {
+        get => m_isUrgent;
+        set => RaiseWhenSet(ref m_isUrgent, value);
+    }
+
+    public bool IsSecondaryUrgent
+    {
+        get => m_isSecondaryUrgent;
+        set => RaiseWhenSet(ref m_isSecondaryUrgent, value);
+    }
+
+    public CounterDisplayMode Mode
+    {
+        get => m_mode;
+        set => RaiseWhenSet(ref m_mode, value);
+    }
+}