+++ src/library/DIPS.Mobile.UI/Components/Counters/Counter.Properties.cs
@@ -1,0 +1,85 @@
+namespace DIPS.Mobile.UI.Components.Counters;
+
+public partial class Counter
+{
+    public static readonly BindableProperty ValueProperty = BindableProperty.Create(
+        nameof(Value),
+        typeof(int),
+        typeof(Counter),
+        defaultBindingMode: BindingMode.OneWay);
+
+    public static readonly BindableProperty SecondaryValueProperty = BindableProperty.Create(
+        nameof(SecondaryValue),
+        typeof(int),
+        typeof(Counter),
+        defaultBindingMode: BindingMode.OneWay,
+        propertyChanged: ((bindable, _, _) => ((Counter)bindable).OnSecondaryValueChanged()));
+
+    public static readonly BindableProperty IsUrgentProperty = BindableProperty.Create(
+        nameof(IsUrgent),
+        typeof(bool),
+        typeof(Counter),
+        defaultBindingMode: BindingMode.OneWay,
+        propertyChanged: (bindable, _, _) => ((Counter)bindable).OnIsUrgentChanged());
+
+    public static readonly BindableProperty IsSecondaryUrgentProperty = BindableProperty.Create(
+        nameof(IsSecondaryUrgent),
+        typeof(bool),
+        typeof(Counter),
+        defaultBindingMode: BindingMode.OneWay,
+        propertyChanged: (bindable, _, _) => ((Counter)bindable).OnIsSecondaryUrgentChanged());
+
+    public static readonly BindableProperty ModeProperty = BindableProperty.Create(
+        nameof(Mode),
+        typeof(CounterDisplayMode),
+        typeof(Counter),
+        defaultBindingMode: BindingMode.OneWay,
+        propertyChanged: (bindable, _, _) => ((Counter)bindable).OnModeChanged());
+
+    /// <summary>
+    /// The mode determining how the primary and secondary counter should be displayed.
+    /// </summary>
+    public CounterDisplayMode Mode
+    {
+        get => (CounterDisplayMode)GetValue(ModeProperty);
+        set => SetValue(ModeProperty, value);
+    }
+    
+    /// <summary>
+    /// Whether the primary value should have an urgent style.
+    /// </summary>
+    public bool IsUrgent
+    {
+        get => (bool)GetValue(IsUrgentProperty);
+        set => SetValue(IsUrgentProperty, value);
+    }
+    
+    /// <summary>
+    /// Whether the secondary value should have an urgent style.
+    /// </summary>
+    public bool IsSecondaryUrgent
+    {
+        get => (bool)GetValue(IsSecondaryUrgentProperty);
+        set => SetValue(IsSecondaryUrgentProperty, value);
+    }
+    
+    /// <summary>
+    /// The primary counter value.
+    /// </summary>
+    public int Value
+    {
+        get => (int)GetValue(ValueProperty);
+        set => SetValue(ValueProperty, value);
+    }
+    
+    /// <summary>
+    /// The secondary counter value.
+    /// </summary>
+    public int SecondaryValue
+    {
+        get => (int)GetValue(SecondaryValueProperty);
+        set => SetValue(SecondaryValueProperty, value);
+    }
+    
+    
+}